[style]user req joi 파일분리
